requirements.txt
squareup==22.0.0.20220921
configparser
fastapi
uvicorn
aiofiles
jinja2